fix: ruff lint/format errors, mypy type fixes, upgrade CI to Node 24 + latest actions
Agent-Logs-Url: https://github.com/GaryOcean428/vex-agent/sessions/9113d7ad-bdd7-418c-841e-47ff55c1cfae

+++ kernel/coordizer_v2/resonance_bank.py
@@ -422,7 +422,7 @@ class ResonanceBank:
             # Shannon entropy of each simplex point
             safe = np.clip(basin, 1e-12, None)
             total += float(-np.sum(safe * np.log(safe)))
-        return (total / len(self.coordinates)) / max_entropy
+        return float((total / len(self.coordinates)) / max_entropy)
 
     def __len__(self) -> int:
         return len(self.coordinates)
